+++ changes.md
@@ -15,7 +15,8 @@
 * Added db-error.php dropin for displaying a custom message for failed DB connections.
 * Changed DB_HOST to rds.discodonniepresents.com which should be overwritten in hosts to point to local MySQL host.
 * Configured ElasticSearch settings for Fantastic Elastic to be network-delegated and moved menu under Settings.
-* Elasticsearch now takes its credentials from consants - WP_ELASTIC_SECRET_KEY and WP_ELASTIC_SERVICE_INDEX.
+* ElasticSearch now takes its credentials from consants - WP_ELASTIC_SECRET_KEY and WP_ELASTIC_SERVICE_INDEX.
+* Updated ElasticSearch settings to use site-specific options - app:api:url, app:api:key and app:api:secret, delegated by policy.
 
 #### 2.2.1
 * Disabled PageSpeed and NewRelic on admin.